Migrate to sveltekit
