+++ website/docs/package.json
@@ -1,18 +1,17 @@
 {
   "name": "luigi-documentation",
-  "description": "Interactive documentation based on svelte3, sapper and luigi",
+  "description": "Interactive documentation based on svelte3, vite and luigi",
   "private": true,
   "scripts": {
     "start": "run-p dev:*",
-    "dev:sapper": "PORT=4001 sapper dev",
     "dev:config": "npm run build:config -- --watch",
     "dev:luigi": "sirv static/luigi --single --dev --cors --port 4000 --host 0.0.0.0",
-    "debug:sapper": "PORT=4001 node --inspect node_modules/sapper/dist/cli.js dev",
+    "debug:vite": "PORT=4001 node --inspect node_modules/vite/dist/cli.js dev",
     "fetchversions": "curl -H \"Accept: application/vnd.github.v3+json\" https://api.github.com/repos/SAP/luigi/tags?per_page=100 > static/versions.json",
-    "build:config": "npm run fetchversions && node -r esm scripts/generate-docu-sitemap.js && webpack --config webpack-generateConfig.config",
-    "export:sapper": "NODE_ENV=production sapper export --basepath /docu-microfrontend && sh ./scripts/move-export.sh",
+    "build:config": "npm run fetchversions && node -r esm scripts/generate-docu-sitemap.js && svelte --config svelte.config.js",
+    "export:vite": "NODE_ENV=production vite export --basepath /docu-microfrontend && sh ./scripts/move-export.sh",
     "export:serve": "sirv public --single --dev --cors --port 5000 --host 0.0.0.0",
-    "export": "npm run build:config && npm run export:sapper",
+    "export": "npm run build:config && npm run export:vite",
     "mock:develop-plugins": "NODE_ENV=debug node -r esm scripts/markdown.unified-preview.js",
     "mock:develop-plugins:watch": "NODE_ENV=debug nodemon --watch src --watch scripts -e js,md -r esm scripts/markdown.unified-preview.js"
   },
@@ -20,8 +19,7 @@
     "@luigi-project/client": "1.4.0",
     "@luigi-project/core": "1.4.0",
     "algoliasearch": "^4.6.0",
-    "docsearch.js": "2.6.3",
-    "polka": "next"
+    "docsearch.js": "2.6.3"
   },
   "devDependencies": {
     "@agentofuser/rehype-section": "^1.0.5",
@@ -31,9 +29,11 @@
     "@babel/plugin-transform-runtime": "^7.5.5",
     "@babel/preset-env": "^7.4.2",
     "@babel/runtime": "^7.5.5",
+    "@sveltejs/adapter-auto": "^1.0.0-next.67",
+    "@sveltejs/adapter-static": "^1.0.0-next.39",
+    "@sveltejs/kit": "^1.0.0-next.445",
     "babel-loader": "^8.0.4",
     "clean-webpack-plugin": "^1.0.0",
-    "compression": "^1.7.1",
     "copy-webpack-plugin": "^6.2.1",
     "css-loader": "^3.2.0",
     "esm": "^3.2.25",
@@ -47,15 +47,14 @@
     "remark-frontmatter": "^1.3.2",
     "remark-parse": "^8.0.3",
     "remark-rehype": "^8.0.0",
-    "sapper": "^0.27.12",
     "sass": "^1.49.8",
     "sass-loader": "^8.0.0",
-    "sirv-cli": "0.4.4",
     "svelte": "^3.49.0",
     "svelte-loader": "^2.9.0",
     "svelte-preprocess": "^3.2.0",
     "unified": "^8.3.2",
     "webpack": "^4.42.1",
     "webpack-cli": "^3.3.12"
-  }
+  },
+  "type": "module"
 }
