+++ website/docs/src/app.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta
+      name="viewport"
+      content="width=device-width, user-scalable=no, initial-scale=1, maximum-scale=1, minimum-scale=1"
+    />
+    <meta name="theme-color" content="#333333" />
+
+    <link rel="stylesheet" href="global.css" />
+    <!-- <link rel='manifest' href='manifest.json'> -->
+    <!-- <link rel='icon' type='image/png' href='favicon.png'> -->
+
+    <!-- This contains the contents of the <svelte:head> component, if
+	     the current page has one -->
+    %sveltekit.head%
+  </head>
+  <body>
+    <!-- The application will be rendered inside this element,
+		 because `src/client.js` references it -->
+    %sveltekit.body%
+
+    <!-- Sapper creates a <script> tag containing `src/client.js`
+	     and anything else it needs to hydrate the app and
+		 initialise the router -->
+    <script type="text/javascript">
+      let script = document.createElement('script');
+      script.type = 'text/javascript';
+      if (location.port == '4001') {
+        script.src = '/luigi-client/luigi-client.js';
+      } else {
+        script.src = '/docu-microfrontend/luigi-client/luigi-client.js';
+      }
+
+      document.getElementsByTagName('body')[0].appendChild(script);
+    </script>
+  </body>
+</html>
