+++ website/docs/svelte.config.js
@@ -1,0 +1,7 @@
+import adapter from '@sveltejs/adapter-static';
+
+export default {
+  kit: {
+    adapter: adapter()
+  }
+};
